backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
youtube-transcript-api==0.6.2
gunicorn==21.2.0
google-genai>=1.0.0
python-dotenv==1.0.1
groq==0.4.2
yt-dlp==2025.01.26
